% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/factToNumericMatrix.R
\name{factToNumericMatrix}
\alias{factToNumericMatrix}
\title{A function to convert a table of triples into a 2-dimensional numeric matrix.}
\usage{
factToNumericMatrix(T, n.threads = 1)
}
\arguments{
\item{T}{A three column table containing the data to cast into a matrix -- the first column contains row indices, the second column contains column indices and the third column contains the values to be stored at each location in the output matrix.
This can either be a data frame with the first two columns of type character and the third of type numeric, or a numeric three-column matrix.}

\item{n.threads}{The number of threads to use.  You probably don't want to ask for more cores than are available to your R process.
Internally, the package limits the number of threads to floor((number of rows * number of cols)/2).}
}
\description{
This function allows you to e.g. convert a fact table from SQL into a matrix.
It uses multiple threads and compiled C code to be more efficient than native R
alternatives.
}
\examples{
factToNumericMatrix(data.frame(c("a","a","a","b","b","b","c","c","c"), rep(times=3, c("x", "y", "z")), as.numeric(1:9)), 1)
}
\keyword{reshape}

